chore: dedupe postgrest error wrap and team-storage helpers
